+++ src/XmlDocument.php
@@ -116,6 +116,22 @@ class XmlDocument
      *
      * @return string
      */
+    public static function requireString($inputVar)
+    {
+        if (!\is_string($inputVar)) {
+            throw new XmlDocumentException('expected "string"');
+        }
+
+        return $inputVar;
+    }
+
+    /**
+     * @param mixed $inputVar
+     *
+     * @throws \fkooman\SAML\SP\Exception\XmlDocumentException
+     *
+     * @return string
+     */
     public static function requireNonEmptyString($inputVar)
     {
         if (!\is_string($inputVar)) {